Project/backend/requirements.txt
asgiref==3.10.0
black==25.11.0
boto3==1.41.5
botocore==1.41.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
flake8==7.3.0
idna==3.11
isort==7.0.0
jmespath==1.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pipreqs==0.4.13
platformdirs==4.5.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
